+++ common/src/main/java/ho/artisan/azusaconfig/AzusaConfigExpectPlatform.java
@@ -1,6 +1,7 @@
 package ho.artisan.azusaconfig;
 
 import dev.architectury.injectables.annotations.ExpectPlatform;
+import dev.architectury.injectables.annotations.PlatformOnly;
 
 import java.nio.file.Path;
 import java.util.List;
@@ -12,9 +13,10 @@ public class AzusaConfigExpectPlatform {
         throw new AssertionError();
     }
 
-//    @ExpectPlatform
-//    public static <T> List<T> getEntrypoints(String key, Class<T> type) {
-//        // Just throw an error, the content should get replaced at runtime.
-//        throw new AssertionError();
-//    }
+    @PlatformOnly("fabric")
+    @ExpectPlatform
+    public static <T> List<T> getEntrypoints(String key, Class<T> type) {
+        // Just throw an error, the content should get replaced at runtime.
+        throw new AssertionError();
+    }
 }