+++ common/src/main/java/ho/artisan/azusaconfig/AzusaConfigMod.java
@@ -8,8 +8,6 @@ import org.slf4j.MarkerFactory;
 public class AzusaConfigMod {
     public static final String MOD_ID = "azusaconfig";
     public static final String MOD_NAME = "AzusaConfig";
-
-
     public static final Logger LOGGER = LoggerFactory.getLogger(MOD_NAME);
     public static final Marker MARKER = MarkerFactory.getMarker("Config");
 }