+++ forge/src/main/java/ho/artisan/azusaconfig/forge/AzusaConfigModForge.java
@@ -1,8 +1,7 @@
 package ho.artisan.azusaconfig.forge;
 
 import ho.artisan.azusaconfig.AzusaConfigMod;
-//import ho.artisan.azusaconfig.forge.entrypoint.impl.EntrypointStorage;
-import ho.artisan.azusaconfig.impl.config.AzusaConfigImpl;
+import ho.artisan.azusaconfig.impl.AzusaConfigImpl;
 import net.minecraftforge.fml.IExtensionPoint;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
@@ -12,7 +11,6 @@ import net.minecraftforge.network.NetworkConstants;
 public class AzusaConfigModForge {
     public AzusaConfigModForge() {
         ModLoadingContext.get().registerExtensionPoint(IExtensionPoint.DisplayTest.class, () -> new IExtensionPoint.DisplayTest(() -> NetworkConstants.IGNORESERVERONLY, (a, b) -> true));
-//        EntrypointStorage.loadAll();
         AzusaConfigImpl.init();
     }
 }